refactor(app): clean up and fix svg deck map layers PV
closes AUTH-2423 AUTH-2428
