+++ app/src/pages/Desktop/Protocols/ProtocolVisualization/DeckViewOverlay.tsx
@@ -2,6 +2,7 @@ import { useMemo } from 'react'
 
 import {
   ALIGN_CENTER,
+  COLORS,
   CURSOR_POINTER,
   RobotCoordsForeignObject,
 } from '@opentrons/components'
@@ -9,10 +10,11 @@ import {
   FLEX_ROBOT_TYPE,
   getCutoutIdForAddressableArea,
   getDeckDefFromRobotType,
+  getFlexHoverDimensions,
+  getOT2HoverDimensions,
   THERMOCYCLER_MODULE_TYPE,
 } from '@opentrons/shared-data'
 
-import { getFlexHoverDimensions } from './utils'
 import styles from './visualization.module.css'
 
 import type { Dispatch, ReactNode, SetStateAction } from 'react'
@@ -28,17 +30,14 @@ interface SlotOverlayProps {
   slotId: DeckSlotId
   slotPosition: CoordinateTuple | null
   slotFillColor: string
-  children: ReactNode
-  robotType: RobotType
   invariantContext: InvariantContext
   robotState: RobotState
   setSelectedSlot: Dispatch<SetStateAction<string | null>>
   setHoveredSlot: Dispatch<SetStateAction<string | null>>
-  opacity?: number
+  robotType: RobotType
+  hover: string | null
+  children?: ReactNode
 }
-
-const TC_X_OFFSET = 20
-const TC_Y_OFFSET = 68
 
 export function DeckViewOverlay(props: SlotOverlayProps): JSX.Element | null {
   const {
@@ -51,7 +50,7 @@ export function DeckViewOverlay(props: SlotOverlayProps): JSX.Element | null {
     robotState,
     setSelectedSlot,
     setHoveredSlot,
-    opacity = 1,
+    hover,
   } = props
   const { stagingAreaEntities, moduleEntities } = invariantContext
   const { modules } = robotState
@@ -76,48 +75,101 @@ export function DeckViewOverlay(props: SlotOverlayProps): JSX.Element | null {
   if (slotPosition === null || (hasTCOnSlot && tcSlots.includes(slotId))) {
     return null
   }
+  const hoverOpacity = hover != null && hover === slotId ? 0.9 : 0
 
-  // TODO: extend for Ot-2
-  const { width, x, y } = getFlexHoverDimensions(
-    stagingAreaLocations,
-    cutoutId,
-    slotId,
-    hasTCOnSlot != null,
-    slotPosition
-  )
+  if (robotType === FLEX_ROBOT_TYPE) {
+    const { width, x, y, height } = getFlexHoverDimensions(
+      stagingAreaLocations,
+      cutoutId,
+      slotId,
+      hasTCOnSlot != null,
+      slotPosition
+    )
 
-  return (
-    <RobotCoordsForeignObject
-      key={`${robotType.toLowerCase()}_slotOverlay`}
-      width={width}
-      height="6%"
-      x={x + (hasTCOnSlot ? TC_X_OFFSET : 0)}
-      y={y - (hasTCOnSlot ? TC_Y_OFFSET : 0)}
-      flexProps={{ flex: '1' }}
-      foreignObjectProps={{
-        opacity: opacity,
-        flex: '1',
-        cursor: CURSOR_POINTER,
-        textAlign: ALIGN_CENTER,
-      }}
-      foreignObjectEvents={{
-        onClick: () => {
-          setSelectedSlot(slotId)
-        },
-        onMouseEnter: () => {
-          setHoveredSlot(slotId)
-        },
-        onMouseLeave: () => {
-          setHoveredSlot(null)
-        },
-      }}
-    >
-      <div
-        className={styles.deck_overlay}
-        style={{ backgroundColor: slotFillColor }}
+    return (
+      <RobotCoordsForeignObject
+        key={`${robotType.toLowerCase()}_slotOverlay`}
+        width={width}
+        height={height}
+        x={x}
+        y={y}
+        flexProps={{ flex: '1' }}
+        foreignObjectProps={{
+          opacity: hoverOpacity,
+          flex: '1',
+          cursor: CURSOR_POINTER,
+          textAlign: ALIGN_CENTER,
+          border: `3px solid ${COLORS.purple50}`,
+          borderRadius: '6px',
+        }}
+        foreignObjectEvents={{
+          onClick: () => {
+            setSelectedSlot(slotId)
+          },
+          onMouseEnter: () => {
+            setHoveredSlot(slotId)
+          },
+          onMouseLeave: () => {
+            setHoveredSlot(null)
+          },
+        }}
       >
-        {children}
-      </div>
-    </RobotCoordsForeignObject>
-  )
+        {children != null ? (
+          <div className={styles.deck_overlay}>
+            <div
+              className={styles.text_background}
+              style={{ backgroundColor: slotFillColor }}
+            >
+              {children}
+            </div>
+          </div>
+        ) : null}
+      </RobotCoordsForeignObject>
+    )
+  } else {
+    const { width, x, y, height } = getOT2HoverDimensions(
+      hasTCOnSlot != null,
+      slotPosition
+    )
+
+    return (
+      <RobotCoordsForeignObject
+        key="ot2_hover"
+        width={width}
+        height={height}
+        x={x}
+        y={y}
+        flexProps={{ flex: '1' }}
+        foreignObjectProps={{
+          opacity: hoverOpacity,
+          flex: '1',
+          cursor: CURSOR_POINTER,
+          border: `3px solid ${COLORS.purple50}`,
+          borderRadius: '6px',
+        }}
+        foreignObjectEvents={{
+          onClick: () => {
+            setSelectedSlot(slotId)
+          },
+          onMouseEnter: () => {
+            setHoveredSlot(slotId)
+          },
+          onMouseLeave: () => {
+            setHoveredSlot(null)
+          },
+        }}
+      >
+        {children != null ? (
+          <div className={styles.deck_overlay}>
+            <div
+              className={styles.text_background}
+              style={{ backgroundColor: slotFillColor }}
+            >
+              {children}
+            </div>
+          </div>
+        ) : null}
+      </RobotCoordsForeignObject>
+    )
+  }
 }
