+++ app/src/pages/Desktop/Protocols/ProtocolVisualization/visualization.module.css
@@ -88,14 +88,13 @@
   position: relative;
   height: var(--spacing-4);
   border-radius: var(--border-radius-2);
-  background:
-    linear-gradient(
-      to right,
-      var(--blue-50) 0%,
-      var(--blue-50) var(--progress, 0%),
-      #ccc var(--progress, 0%),
-      #ccc 100%
-    );
+  background: linear-gradient(
+    to right,
+    var(--blue-50) 0%,
+    var(--blue-50) var(--progress, 0%),
+    #ccc var(--progress, 0%),
+    #ccc 100%
+  );
 }
 
 /* Hide thumb initially */
@@ -355,10 +354,19 @@
   display: flex;
   width: 100%;
   height: 100%;
-  align-items: center;
   justify-content: center;
+  align-items: flex-end;
   border-radius: var(--border-radius-4);
   gap: var(--spacing-8);
+  background-color: transparent;
+}
+
+.text_background {
+  display: flex;
+  justify-content: center;
+  width: 100%;
+  background-color: var(--slot-fill-color);
+  border-radius: 0 0 var(--border-radius-4) var(--border-radius-4);
 }
 
 .controls_overflow_menu {
