+++ shared-data/js/helpers/index.ts
@@ -64,6 +64,7 @@ export * from './getModuleDeckLabel'
 export * from './deckConfig'
 export * from './testHelpers'
 export * from './symbolicPositionHelpers'
+export * from './slotHovers'
 
 export const getLabwareDefIsStandard = (def: LabwareDefinition): boolean =>
   def?.namespace === OPENTRONS_LABWARE_NAMESPACE
